refactor: inputView로 Position 파싱 책임 이전
